safari-600.7-branch/LayoutTests/fast/selectors/querySelector-multiple-attribute-match-with-child-backtracking-expected.txt

Test backtracking of matching multiple attribute with a single child selector relation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("[data-foo][data-bar]>li [style]").length is 2
PASS document.querySelectorAll("[data-foo][data-bar]>li [style]")[0].id is "target1"
PASS document.querySelectorAll("[data-foo][data-bar]>li [style]")[1].id is "target2"
PASS successfullyParsed is true

TEST COMPLETE
